refactoring the index.js file into components file architecture
